Continue transition to feature based directory structure. Clean up SignIn, SignUp components. Utilize react-router-dom's layout components
