+++ client/src/features/SignIn/ForgotPassword.jsx
@@ -1,0 +1,9 @@
+export default function ForgotPassword() {
+  return (
+    <div className="text-sm">
+      <a href="#" className="font-medium text-indigo-600 hover:text-indigo-500">
+        Forgot your password?
+      </a>
+    </div>
+  );
+}
